Fix compilation of the flare example.
Former-commit-id: 00f002aa143973af442d66af6da091b9a4c6987d

+++ example/flares/src/main.cpp
@@ -155,9 +155,9 @@ main(int argc, char** argv)
         camera->component<Transform>()->matrix()->lookAt(
             lookAt,
             Vector3::create(
-                lookAt->x() + distance * std::cosf(yaw) * sinf(pitch),
-                lookAt->y() + distance * std::cosf(pitch),
-                lookAt->z() + distance * std::sinf(yaw) * sinf(pitch)
+                lookAt->x() + distance * std::cos(yaw) * std::sin(pitch),
+                lookAt->y() + distance * std::cos(pitch),
+                lookAt->z() + distance * std::sin(yaw) * std::sin(pitch)
             )
         );
